rag-chatbot/requirements.txt
# RAG Chatbot Application Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.21.0
openai>=1.12.0
azure-cosmos>=4.5.0
requests>=2.28.0

# Machine Learning
scikit-learn>=1.3.0

# Environment management
python-dotenv>=1.0.0

# Optional for enhanced functionality
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# For development and testing
pytest>=7.0.0
pytest-cov>=4.0.0